seperate MarketCardsOffer form others
